requirements.txt
# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Arabic Text Processing
arabic-reshaper==3.0.0
python-bidi==0.4.2

# Database
sqlalchemy==2.0.23

# Data Export
openpyxl==3.1.2
xlsxwriter==3.1.9

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2

# Browser Driver Manager
webdriver-manager==4.0.1

# Testing (optional)
pytest==7.4.3